requirements.txt
Django
djangorestframework
Markdown
django-filter
psycopg2-binary
